+++ src/verb/builtin.rs
@@ -38,7 +38,11 @@ pub fn builtin_verbs() -> Vec<Verb> {
         )
         .unwrap()
         .with_shortcut("cpp"),
-        Verb::internal(focus), // hardcoded Enter
+        // :focus is also hardcoded on Enter on directories
+        // but ctrl-f is useful for focusing on a file's parent
+        // (and keep the filter)
+        Verb::internal(focus)
+            .with_control_key('f'),
         Verb::internal(help).with_key(F1).with_shortcut("?"),
         Verb::internal(line_down).with_key(DOWN),
         Verb::internal(line_up).with_key(UP),