+++ src/browser/browser_state.rs
@@ -665,6 +665,7 @@ impl AppState for BrowserState {
             tree: &self.displayed_tree(),
             skin: &panel_skin.styles,
             cols: &con.cols,
+            show_selection_mark: con.show_selection_mark,
             ext_colors: &con.ext_colors,
             area,
             in_app: true,