+++ src/conf/conf.rs
@@ -35,6 +35,7 @@ pub struct Conf {
     pub search_modes: SearchModeMap,
     pub disable_mouse_capture: bool,
     pub cols_order: Option<Cols>,
+    pub show_selection_mark: Option<bool>,
     pub ext_colors: ExtColorMap,
     pub syntax_theme: Option<String>,
     pub true_colors: Option<bool>,
@@ -242,6 +243,10 @@ impl Conf {
         // true_colors ?
         if let Some(b) = bool_field(&root, "true_colors") {
             self.true_colors = Some(b);
+        }
+        // show selection mark
+        if let Some(b) = bool_field(&root, "show_selection_mark") {
+            self.show_selection_mark = Some(b);
         }
 
         Ok(())