+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### next version
 - completion of the "client-server" feature (see client-server.md)
 - removing a filter using the esc key on a tree tries to keep the selection
+- :focus now has a shortcut for when a file is selected too: ctrl-f
 
 <a name="v0.20.0"></a>
 ### v0.20.0 - 2020-08-16